LayoutTests/imported/w3c/web-platform-tests/selection/onselectstart-on-key-in-contenteditable_preventDefault=yes-expected.txt

abc
def
ghi
jkl
mno
pqr
stu
vwx
yz0
123
456
789

PASS Shift+ArrowLeft when <div contenteditable>[]abc<br>
PASS Shift+ArrowLeft when <div contenteditable>a[]bc<br>
PASS Shift+ArrowRight when <div contenteditable>a[]bc<br>
PASS Shift+ArrowRight when 789[]<br></div>
PASS Shift+ArrowUp when <div contenteditable>[]abc<br>
PASS Shift+ArrowUp when <div contenteditable>a[]bc<br>
PASS Shift+ArrowDown when <div contenteditable>a[]bc<br>
PASS Shift+ArrowDown when 789[]<br></div>
PASS Shift+ArrowDown when 78[]9<br></div>
PASS Shift+ArrowUp when 78[]9<br></div>
PASS Shift+PageUp when <div contenteditable>[]abc<br>
FAIL Shift+PageUp when <div contenteditable>a[]bc<br> assert_equals: Shift+PageUp when <div contenteditable>a[]bc<br>: selection range should not be changed if preventDefault() of "selectstart" event is called expected "(#text \"abc\", 1)(collapsed)" but got "(#text \"abc\", 0) - (#text \"abc\", 1)"
FAIL Shift+PageDown when <div contenteditable>a[]bc<br> assert_equals: Shift+PageDown when <div contenteditable>a[]bc<br>: selection range should not be changed if preventDefault() of "selectstart" event is called expected "(#text \"abc\", 1)(collapsed)" but got "(#text \"abc\", 1) - (#text \"jkl\", 1)"
PASS Shift+PageDown when 789[]<br></div>
FAIL Shift+PageDown when 78[]9<br></div> assert_equals: Shift+PageDown when 78[]9<br></div>: selection range should not be changed if preventDefault() of "selectstart" event is called expected "(#text \"789\", 2)(collapsed)" but got "(#text \"789\", 2) - (#text \"789\", 3)"
FAIL Shift+PageUp when 78[]9<br></div> assert_equals: Shift+PageUp when 78[]9<br></div>: selection range should not be changed if preventDefault() of "selectstart" event is called expected "(#text \"789\", 2)(collapsed)" but got "(#text \"yz0\", 2) - (#text \"789\", 2)"
FAIL Shift+Home when <div contenteditable>[]abc<br> assert_equals: Shift+Home when <div contenteditable>[]abc<br>: selection range should not be changed if preventDefault() of "selectstart" event is called expected "(<div contenteditable=\"\">, 0)(collapsed)" but got "(#text \"abc\", 0)(collapsed)"
PASS Shift+Home when <div contenteditable>a[]bc<br>
PASS Shift+End when <div contenteditable>a[]bc<br>
PASS Shift+End when 789[]<br></div>
PASS Select All when 789[]<br></div>
PASS Shift+ArrowLeft when d[e]f<br>
PASS Shift+ArrowRight when d[e]f<br>
PASS Shift+ArrowUp when d[e]f<br>
PASS Shift+ArrowDown when d[e]f<br>
PASS Shift+PageDown when d[e]f<br>
PASS Shift+PageUp when d[e]f<br>
PASS Shift+Home when d[e]f<br>
PASS Shift+End when d[e]f<br>
PASS Select All when d[e]f<br>
